Fix(RecipeDetail): Correcion de errores en Store y Update Request y modelo

+++ app/Models/RecipeDetail.php
@@ -20,7 +20,7 @@ class RecipeDetail extends Model
     ];
 
     protected $casts = [
-        'quantity' => 'integer'
+        'quantity' => 'decimal'
     ];
 
     public function recipe(): BelongsTo
