+++ database/migrations/2026_06_30_014830_create_recipe_details_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('recipe_details', function (Blueprint $table) {
             $table->id();
-            $table->integer('quantity');
+            $table->decimal('quantity', 8, 2);
             $table->string('instruction');
 
             $table->foreignId('recipe_id')
